refactor: 認証の Supabase 基盤を infrastructure ディレクトリへ集約
supabase-client / supabase-result は Supabase への依存を閉じ込める技術基盤のため、
リポジトリの DDD 慣習(article/infrastructure 等)に倣い infrastructure/ に集約する。
用途ディレクトリ(oauth 等)とは層が異なることをディレクトリ名で明示する。
