+++ application/packages/authentication/src/oauth/client.ts
@@ -5,8 +5,8 @@ import {
   type AuthClientConfig,
   createBackendClient,
   type SupabaseAuthClient,
-} from '../supabase-client'
-import { callSupabase } from '../supabase-result'
+} from '../infrastructure/supabase-client'
+import { callSupabase } from '../infrastructure/supabase-result'
 import type { OAuthProvider } from './provider'
 
 export class OAuthClient {
