+++ application/packages/authentication/src/passkey/client.ts
@@ -10,8 +10,8 @@ import {
   type AuthClientConfig,
   createBackendClient,
   type SupabaseAuthClient,
-} from '../supabase-client'
-import { callSupabase } from '../supabase-result'
+} from '../infrastructure/supabase-client'
+import { callSupabase } from '../infrastructure/supabase-result'
 
 type PasskeyApi = SupabaseAuthClient['auth']['passkey']
 type VerifyRegistrationParams = Parameters<PasskeyApi['verifyRegistration']>[0]
