+++ application/packages/authentication/src/password/client.ts
@@ -10,8 +10,8 @@ import {
   type AuthClientConfig,
   createBackendClient,
   type SupabaseAuthClient,
-} from '../supabase-client'
-import { callSupabase } from '../supabase-result'
+} from '../infrastructure/supabase-client'
+import { callSupabase } from '../infrastructure/supabase-result'
 
 interface PasswordCredentials {
   email: string
