+++ application/packages/authentication/src/session/client.ts
@@ -4,8 +4,8 @@ import {
   type AuthClientConfig,
   createBackendClient,
   type SupabaseAuthClient,
-} from '../supabase-client'
-import { callSupabase } from '../supabase-result'
+} from '../infrastructure/supabase-client'
+import { callSupabase } from '../infrastructure/supabase-result'
 
 export class SessionClient {
   private readonly client: SupabaseAuthClient
